cleancode整改

Created-by: ninihihi
Commit-by: zouyanlong
Merged-by: turing_project1
Description: ## 描述
cleancode 8.5
## checklist
<!--  [x] 表示选中 -->
- [x] 是否通过本地IDE对代码进行静态检查
- [x] 是否通过本地IDE对代码进行格式化处理
- [x] 是否进行空指针校验
- [x] 是否进行返回值校验
- [x] 是否正确释放new/malloc申请的内存
- [x] 是否充分考虑接口的异常场景
- [x] 是否正确记录错误日志

See merge request: cann/ascend-transformer-boost!1781

+++ src/kernels/mixkernels/multi_latent_attention/tiling/mla_tiling_dependency.cpp
@@ -173,9 +173,10 @@ Status GetNdMLATiling(const MLAInfo &mmInfo, uint32_t &blockDim, uint32_t *tilin
         tilingParam[tilingOffset] = static_cast<uint32_t>(qSeqLen);
         tilingParam[tilingOffset + 1] = static_cast<uint32_t>(kvSeqlen);
         GetAddrOffsetMLA(tilingParam, addrOffsets, tilingOffset);
-        uint64_t addressQffset = static_cast<uint64_t>(mmInfo.numHeads) * qSeqLen;
-        uint64_t addressOffset = static_cast<uint64_t>(mmInfo.numHeads) * mmInfo.embeddingSize * qSeqLen;
-        uint64_t addressMaskOffset = static_cast<uint64_t>(qSeqLen) * maxKVseqlen;
+        uint64_t addressQffset = static_cast<uint64_t>(mmInfo.numHeads) * static_cast<uint64_t>(qSeqLen);
+        uint64_t addressOffset = static_cast<uint64_t>(mmInfo.numHeads) * static_cast<uint64_t>(mmInfo.embeddingSize) *
+                                 static_cast<uint64_t>(qSeqLen);
+        uint64_t addressMaskOffset = static_cast<uint64_t>(qSeqLen) * static_cast<uint64_t>(maxKVseqlen);
         if (mmInfo.maskType == OpParam::MLA::MASK_TYPE_MASK_FREE) {
             addressMaskOffset = mmInfo.maskStride;
         }
